requirements.txt
flask=1.1.1
flask_socketio=4.2.1
flask_pymongo=2.3.0